Restore admin dashboard layout and archive alternate version
